# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HiveTrust API",
    "version": "1.0.0",
    "description": "KYA Identity Verification, Trust Scoring & Parametric Insurance for AI Agents in the A2A Economy.\n\nAll endpoints require an `X-API-Key` header unless noted. Costs are billed per call and reflected in your operator account.",
    "contact": {
      "name": "HiveTrust Support",
      "url": "https://hivetrust.hiveagentiq.com",
      "email": "support@hiveagentiq.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://hivetrust.hiveagentiq.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "security": [
    { "ApiKeyAuth": [] }
  ],
...
  "paths": {
...
    "/v1/verify_agent_risk": {
      "get": {
        "tags": ["Trust Score"],
        "summary": "Quick risk check",
        "description": "Returns a binary `clear` or `block` decision optimised for payment-processor latency (< 50ms). Cost: $0.01.",
        "operationId": "verifyAgentRisk",
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "min_score",
            "in": "query",
            "schema": { "type": "integer", "minimum": 0, "maximum": 1000 }
          }
        ],
        "responses": {
          "200": {
            "description": "Risk decision",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RiskCheckResponse" }
              }
            }
          }
        }
      }
    },
...
    "/mcp": {
      "post": {
        "tags": ["System"],
        "summary": "MCP JSON-RPC 2.0 endpoint",
        "description": "Model Context Protocol endpoint. Supports `tools/list` and `tools/call` methods.",
        "operationId": "mcpEndpoint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MCPRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MCP response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MCPResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key obtained from the HiveTrust dashboard or via the seed script for testing."
      }
    },
...
    "schemas": {
...
      "RiskCheckResponse": {
        "type": "object",
        "properties": {
          "agent_id": { "type": "string", "format": "uuid" },
          "decision": { "type": "string", "enum": ["clear", "block"] },
          "score": { "type": "integer" },
          "tier": { "type": "string" },
          "latency_ms": { "type": "number" }
        }
      },
...
      "MCPRequest": {
        "type": "object",
        "required": ["jsonrpc", "method"],
        "properties": {
          "jsonrpc": { "type": "string", "enum": ["2.0"] },
          "id": { "oneOf": [{ "type": "string" }, { "type": "integer" }] },
          "method": { "type": "string", "enum": ["tools/list", "tools/call"] },
          "params": { "type": "object" }
        }
      },
      "MCPResponse": {
        "type": "object",
        "properties": {
          "jsonrpc": { "type": "string", "enum": ["2.0"] },
          "id": { "oneOf": [{ "type": "string" }, { "type": "integer" }] },
          "result": { "type": "object" },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "integer" },
              "message": { "type": "string" }
            }
          }
        }
      },
...
    }
  }
}